feat: Phase 21 — Claude-side query trigger + AST injection fix + decision delete
Phase 21 (Claude-Side Query Trigger):
- Extract shared query module (domain/query.rs) from user_prompt_submit
- Add `base context <text>` subcommand — same matching engine as hooks
- Add `base context --list` — shows all triggerable domains + query files
- Bypass dedup on explicit calls, register hash for hook-side suppression
- Session-start cheat-sheet (<base-context-triggers> block)

AST injection fix (Phase 19 residual):
- file_map_compact used exact match (=) instead of CONTAINS — never matched
- Absolute paths from hooks not stripped to workspace-relative — FILTER failed
- Both fixed; AST entity maps now inject on Read/Edit of source files

Additional:
- `base decision delete --keyword "..."` — CRUD gap, shows matches then deletes
- SPARQL query template (_template.sparql) with full authoring contract
- offer-capabilities.sparql rewritten from static to live graph query
