+++ src/hook/user_prompt_submit.rs
@@ -4,9 +4,9 @@ use anyhow::Result;
 use oxigraph::model::TermRef;
 
 use crate::config::BaseConfig;
-use crate::crud;
 use crate::domain;
 use crate::domain::matcher::match_domains;
+use crate::domain::query::{query_domain_from_graph, resolve_and_run_query, format_toml_rules};
 use crate::domain::session::{rules_hash, Bracket, SessionState};
 
 pub fn handle(config: &BaseConfig, cwd: &Path, event: &serde_json::Value) -> Result<super::HookEventData> {
@@ -316,240 +316,6 @@ pub fn format_devmode_block(
 }
 
 // ─── Graph-backed injection ─────────────────────────────────
-
-/// Query a domain's rules and 1-hop neighborhood from the graph.
-/// Returns (rules_text, neighborhood_text). Falls back to TOML if graph query fails.
-fn query_domain_from_graph(
-    store: &oxigraph::store::Store,
-    config: &BaseConfig,
-    domain_def: &domain::DomainDef,
-) -> (String, String) {
-    let ns = &config.namespace;
-    let p = &ns.prefix;
-    let domain_slug = crud::slugify(&domain_def.name);
-    let domain_iri = crud::build_iri(ns, "domain", &domain_slug);
-    let pfx = crud::prefixes(ns);
-
-    // Query 1: Get rules ordered by priority
-    let rules_sparql = format!(
-        "{pfx}\n\
-         SELECT ?text WHERE {{\n\
-           GRAPH ?g {{\n\
-             <{domain_iri}> {p}:hasRule ?rule .\n\
-             ?rule {p}:ruleText ?text .\n\
-             OPTIONAL {{ ?rule {p}:priority ?pri }}\n\
-           }}\n\
-         }}\n\
-         ORDER BY ?pri"
-    );
-
-    let rules_text = match crate::store::query(store, &rules_sparql) {
-        Ok(oxigraph::sparql::QueryResults::Solutions(solutions)) => {
-            let rules: Vec<String> = solutions
-                .filter_map(|r| r.ok())
-                .filter_map(|row| {
-                    row.get("text").map(|t| match t.into() {
-                        TermRef::Literal(l) => l.value().to_string(),
-                        _ => String::new(),
-                    })
-                })
-                .filter(|s| !s.is_empty())
-                .collect();
-
-            if rules.is_empty() {
-                // Fallback to TOML rules
-                format_toml_rules(domain_def)
-            } else {
-                let mut out = format!("[DOMAIN: {}]\n", domain_def.name);
-                for (i, rule) in rules.iter().enumerate() {
-                    out.push_str(&format!("  {i}. {rule}\n"));
-                }
-                out
-            }
-        }
-        _ => format_toml_rules(domain_def),
-    };
-
-    // Query 2: 1-hop neighborhood (decisions linked to this domain, projects with hasDomain)
-    let neighborhood_sparql = format!(
-        "{pfx}\n\
-         SELECT ?name ?type WHERE {{\n\
-           GRAPH ?g {{\n\
-             {{\n\
-               <{domain_iri}> {p}:hasDecision ?related .\n\
-               ?related {p}:name ?name .\n\
-               BIND({p}:Decision AS ?type)\n\
-             }} UNION {{\n\
-               ?related {p}:hasDomain <{domain_iri}> ;\n\
-                 a {p}:Project ;\n\
-                 {p}:name ?name .\n\
-               BIND({p}:Project AS ?type)\n\
-             }}\n\
-           }}\n\
-         }}\n\
-         ORDER BY ?type ?name"
-    );
-
-    let neighborhood_text = match crate::store::query(store, &neighborhood_sparql) {
-        Ok(oxigraph::sparql::QueryResults::Solutions(solutions)) => {
-            let neighbors: Vec<(String, String)> = solutions
-                .filter_map(|r| r.ok())
-                .filter_map(|row| {
-                    let name = row.get("name").map(|t| match t.into() {
-                        TermRef::Literal(l) => l.value().to_string(),
-                        _ => String::new(),
-                    })?;
-                    let type_label = row.get("type").map(|t| crud::term_display(t.into()))?;
-                    if name.is_empty() {
-                        None
-                    } else {
-                        Some((type_label, name))
-                    }
-                })
-                .collect();
-
-            if neighbors.is_empty() {
-                String::new()
-            } else {
-                let mut out = format!("[{} CONTEXT]\n", domain_def.name);
-                for (type_label, name) in &neighbors {
-                    out.push_str(&format!("  - {type_label}: {name}\n"));
-                }
-                out
-            }
-        }
-        _ => String::new(),
-    };
-
-    (rules_text, neighborhood_text)
-}
-
-// ─── Query-triggered injection ─────────────────────────────
-
-/// Resolve a query name to a `.sparql` file, read it, run it, format results.
-/// Resolution: workspace `.base/queries/{name}.sparql` → global `~/.base-gbl/queries/{name}.sparql`.
-pub fn resolve_and_run_query(
-    store: &oxigraph::store::Store,
-    config: &BaseConfig,
-    cwd: &Path,
-    query_name: &str,
-    format: &str,
-    domain_name: &str,
-) -> String {
-    let filename = format!("{query_name}.sparql");
-
-    // Tier 1: workspace
-    let sparql_content = crate::config::find_workspace_base(cwd)
-        .and_then(|base| std::fs::read_to_string(base.join("queries").join(&filename)).ok())
-        // Tier 2: global
-        .or_else(|| {
-            dirs::home_dir().and_then(|home| {
-                std::fs::read_to_string(home.join(".base-gbl").join("queries").join(&filename)).ok()
-            })
-        });
-
-    let sparql_raw = match sparql_content {
-        Some(s) => s,
-        None => {
-            eprintln!("base: query file not found: queries/{filename}");
-            return String::new();
-        }
-    };
-
-    // Prefix substitution (same pattern as queries.toml)
-    let p = &config.namespace.prefix;
-    let u = &config.namespace.uri;
-    let sparql = sparql_raw
-        .replace("{{prefix}}", p)
-        .replace("{{uri}}", u);
-
-    match crate::store::query(store, &sparql) {
-        Ok(oxigraph::sparql::QueryResults::Solutions(solutions)) => {
-            let rows: Vec<_> = solutions.filter_map(|r| r.ok()).collect();
-            if rows.is_empty() {
-                return String::new();
-            }
-
-            let mut out = format!("<base-query name=\"{query_name}\" domain=\"{domain_name}\">\n");
-
-            let known_vars = ["label", "name", "text", "detail", "type", "status", "value", "count", "created"];
-
-            match format {
-                "table" => {
-                    if let Some(first) = rows.first() {
-                        let vars: Vec<&str> = known_vars.iter()
-                            .filter(|v| first.get(**v).is_some())
-                            .copied()
-                            .collect();
-
-                        if !vars.is_empty() {
-                            out.push_str(&format!("| {} |\n", vars.join(" | ")));
-                            out.push_str(&format!("|{}|\n", vars.iter().map(|_| "---").collect::<Vec<_>>().join("|")));
-                            for row in &rows {
-                                let vals: Vec<String> = vars.iter()
-                                    .map(|v| row.get(*v).map(|t| crud::term_display(t.into())).unwrap_or_default())
-                                    .collect();
-                                out.push_str(&format!("| {} |\n", vals.join(" | ")));
-                            }
-                        }
-                    }
-                }
-                "prose" => {
-                    for row in &rows {
-                        for var in &known_vars[..7] {
-                            if let Some(term) = row.get(*var) {
-                                let val = crud::term_display(term.into());
-                                if !val.is_empty() {
-                                    out.push_str(&format!("{var}: {val}\n"));
-                                }
-                            }
-                        }
-                        out.push('\n');
-                    }
-                }
-                _ => {
-                    // Default: "list"
-                    for row in &rows {
-                        let primary = row.get("label")
-                            .or_else(|| row.get("name"))
-                            .or_else(|| row.get("text"))
-                            .map(|t| crud::term_display(t.into()))
-                            .unwrap_or_default();
-                        let detail = row.get("detail")
-                            .or_else(|| row.get("value"))
-                            .map(|t| crud::term_display(t.into()));
-
-                        if let Some(d) = detail {
-                            out.push_str(&format!("- {primary}: {d}\n"));
-                        } else {
-                            out.push_str(&format!("- {primary}\n"));
-                        }
-                    }
-                }
-            }
-
-            out.push_str("</base-query>\n");
-            out
-        }
-        Ok(_) => String::new(),
-        Err(e) => {
-            eprintln!("base: query '{query_name}' failed: {e}");
-            String::new()
-        }
-    }
-}
-
-/// Format rules directly from the DomainDef struct (TOML fallback).
-fn format_toml_rules(domain_def: &domain::DomainDef) -> String {
-    if domain_def.rules.is_empty() {
-        return String::new();
-    }
-    let mut out = format!("[DOMAIN: {}]\n", domain_def.name);
-    for (i, rule) in domain_def.rules.iter().enumerate() {
-        out.push_str(&format!("  {i}. {rule}\n"));
-    }
-    out
-}
 
 // ─── Auto-sync ──────────────────────────────────────────────
